Somewhat silly rebuild for local model usage
