+++ .example.env
@@ -1,4 +1,9 @@
-# A .env file should exist at the root level of this generator environment and include these 12 values
+# A .env file should exist at the root level of this generator environment and include these values
+
+#if MODEL_TYPE is local, you only need to set LOCAL_MODEL_NAME
+#if MODEL_TYPE is azure_openai then set all AZURE_OPENAI settings
+MODEL_TYPE=local
+LOCAL_MODEL_NAME=phi3.5
 
 AZURE_OPENAI_TEXT_ENDPOINT_KEY=youropenapikey
 AZURE_OPENAI_TEXT_ENDPOINT=https://yourazureopenurl.openai.azure.com/
