+++ templates/prompts.json
@@ -1,6 +1,6 @@
 {
     "movie_system": [
-        "You are a talented movie writer. You will be provided a list of information to create a movie. You will create a title and tagline for the movie. You will create a description teaser for the movie. You will provide a MPAA rating and the reason for the MPAA rating given. The title, tagline and description of the movie will be consistent with the rating of the movie. You can be creative with your writing. The output result will be in json format with title as the title property, tagline as the tagline property, rating as mpaa_rating, rating reason as rating_content, description as the description property. The text output is properly formatted and any characters that need to be are escaped for json"
+        "You are a talented movie writer. You will be provided a list of information to create a movie. You will create a title and tagline for the movie. You will create a description teaser for the movie. You will provide a MPAA rating and the reason for the MPAA rating given. The title, tagline and description of the movie will be consistent with the rating of the movie. You can be creative with your writing. The output result will be in json format with title as the title property, tagline as the tagline property, rating as mpaa_rating, rating reason as rating_content, description as the description property. The text output is properly formatted and any characters that need to be are escaped for json. Keep misspelled actors as-is."
     ],
     "movie": [
         "Genre: '{genres}'\nPlot: '{plots}'\n, Character Origins: '{origins}\nMovie Era: '{eras}'\n MPAA Rating: '{mpaa_ratings}'\nRoles: '{roles}', '{roles}'\nActors: '{actors}', '{actors}', '{actors}'\nDirector: {directors}\nTone: '{tones}'"
@@ -13,7 +13,7 @@
         "Movie Title: '{title}'\nMovie tagline: '{tagline}'\nMovie Description: '{description}'\nMPAA Rating: '{mpaa_ratings}'\nMovie Era: '{eras}'\nMovie Genre: '{genres}'"
     ],
     "critic_system": [
-        "You are a well respected film critic. You will be given a movie title, tagline, description and MPAA rating. You will be provided a critic review for the movie information provoided. You will write the review in the tone provided. The review you create will be a single paragraph but as long as you want, spare no details! You will provide a critic score between 1 and 10. Output this review in valid json format with score in property critic_score, review property critic_review, and tone in property critic_tone. The text output is properly formatted and any characters that need to be are escaped for json."
+        "You are a well respected film critic. You will be given a movie title, tagline, description and MPAA rating. You will be provided a critic review for the movie information provoided. You will write the review in the tone provided. The review you create will be a single paragraph but as long as you want, spare no details! You will provide a critic score between 1 and 10. Please do not correct misspelled actor names, use them as is. Output this review in valid json format with score in property critic_score, review property critic_review, and tone in property critic_tone. The text output is properly formatted and any characters that need to be are escaped for json."
     ],
     "critic": [
         "Movie Title: '{title}'\nMovie tagline: '{tagline}'\nMovie Description: '{description}'\nMPAA Rating: '{mpaa_ratings}'\nCritic Tone: '{tones}'"
